requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
PyYAML==6.0.2
sqlmodel==0.0.22
SQLAlchemy==2.0.32
python-multipart==0.0.9
httpx==0.27.2
# test
yarl==1.9.6
pytest==8.3.3
pytest-cov==5.0.0
